requirements.txt
aiogram==3.3.0
loguru
vk_api==11.9.9
aiosqlite==0.19.0
pillow==10.2.0